+++ search/helpers.py
@@ -6,7 +6,7 @@ from django.conf import settings
 from mohawk import Sender
 from raven.contrib.django.raven_compat.models import client
 
-from activitystream import serializers
+from search import serializers
 
 RESULTS_PER_PAGE = 10
 
@@ -124,6 +124,30 @@ def format_query(query, page):
                             'query': 'Event',
                             'boost': 10000
                         }
+                    }},
+                    {'match': {
+                        'type': {
+                            'query': 'dit:Article',
+                            'boost': 10000
+                        }
+                    }},
+                    {'match': {
+                        'type': {
+                            'query': 'dit:Market',
+                            'boost': 10000
+                        }
+                    }},
+                    {'match': {
+                        'type': {
+                            'query': 'dit:Service',
+                            'boost': 20000
+                        }
+                    }},
+                    {'match': {
+                        'type': {
+                            'query': 'dit:Event',
+                            'boost': 10000
+                        }
                     }}
                 ],
                 'filter': [
@@ -133,7 +157,12 @@ def format_query(query, page):
                             'Opportunity',
                             'Market',
                             'Service',
-                            'Event'
+                            'Event',
+                            'dit:Article',
+                            'dit:Opportunity',
+                            'dit:Market',
+                            'dit:Service',
+                            'dit:Event'
                         ]
                     }}
                 ]